requirements.txt
fastapi
uvicorn[standard]
google-cloud-vision
tensorflow
pillow
numpy
python-multipart